Add new "top" option.
Top takes a numeric argument and only shows that many of the largest queues.

(This kind of messes up the summation, as the queues may change from interval to interval.)

Also renamed "totals" to "sum" becaused I really wanted to use -t for top.

Did a little help message rewriting and minor refactoring.